up: add admin user login and logout API and dto

+++ openjob-server/openjob-server-admin/src/main/java/io/openjob/server/admin/request/notify/NotifyGroupAddRequest.java
@@ -5,9 +5,6 @@ import io.swagger.annotations.ApiModelProperty;
 import lombok.Data;
 
 import java.util.List;
-import javax.validation.constraints.Min;
-import javax.validation.constraints.NotBlank;
-import javax.validation.constraints.NotNull;
 
 /**
  * @author inhere
@@ -34,12 +31,12 @@ public class NotifyGroupAddRequest {
     private Integer deleted;
 
     @ApiModelProperty(value = "Delete time")
-    private Integer deleteTime;
+    private Long deleteTime;
 
     @ApiModelProperty(value = "Update time")
-    private Integer updateTime;
+    private Long updateTime;
 
     @ApiModelProperty(value = "Create time")
-    private Integer createTime;
+    private Long createTime;
 }
