+++ openjob-server/openjob-server-admin/src/main/java/io/openjob/server/admin/request/user/AdminRuleUpdateRequest.java
@@ -5,9 +5,6 @@ import io.swagger.annotations.ApiModelProperty;
 import lombok.Data;
 
 import java.util.List;
-import javax.validation.constraints.Min;
-import javax.validation.constraints.NotBlank;
-import javax.validation.constraints.NotNull;
 
 /**
  * @author inhere
@@ -40,12 +37,12 @@ public class AdminRuleUpdateRequest {
     private Integer deleted;
 
     @ApiModelProperty(value = "Delete time")
-    private Integer deleteTime;
+    private Long deleteTime;
 
     @ApiModelProperty(value = "Update time")
-    private Integer updateTime;
+    private Long updateTime;
 
     @ApiModelProperty(value = "Create time")
-    private Integer createTime;
+    private Long createTime;
 }
